More uint/size_t fixes
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6116 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/GameAttributes.cpp
@@ -302,7 +302,7 @@ void CGameAttributes::OnNumSlotsUpdate(CSynchedJSObjectBase *owner)
 	
 	// Clamp to a preset upper bound.
 	CConfigValue *val=g_ConfigDB.GetValue(CFG_MOD, "max_players");
-	size_t maxPlayers=PS_MAX_PLAYERS;
+	uint maxPlayers=PS_MAX_PLAYERS;
 	if (val)
 		val->GetUnsignedInt(maxPlayers);
 	if (pInstance->m_NumSlots > maxPlayers)
